+++ types/index.d.ts
@@ -1,18 +1,12 @@
 import { TimeoutError, OnCancel } from 'promise.timeout'
 export { TimeoutError, OnCancel }
 
-export interface RetryError extends Error {
+export class RetryError extends Error {
   times: number
   timeout: number
   fn: (...args: any[]) => any
-  erros: Error[]
+  errors: Error[]
 }
-
-/**
- * max 6 parameters. (ts only)
- * R = return
- * P = parameter
- */
 
 export interface RetryOptions {
   times?: number
@@ -20,49 +14,13 @@ export interface RetryOptions {
   onerror?: (err: Error, index: number) => void
 }
 
-declare function pretry<R, P1>(
-  fn: (arg1: P1, onCancel?: OnCancel) => R,
-  options?: RetryOptions
-): (arg1: P1) => R
+type EnsurePromise<T> = T extends Promise<any> ? T : Promise<T>
 
-declare function pretry<R, P1, P2>(
-  fn: (arg1: P1, arg2: P2, onCancel?: OnCancel) => R,
+export default function pretry<T extends unknown[], R>(
+  fn: (...args: [...args: T]) => R,
   options?: RetryOptions
-): (arg1: P1, arg2: P2) => R
-
-declare function pretry<R, P1, P2, P3>(
-  fn: (arg1: P1, arg2: P2, arg3: P3, onCancel?: OnCancel) => R,
+): (...args: T) => EnsurePromise<R>
+export default function pretry<T extends unknown[], R>(
+  fn: (...args: [...args: T, onCancel?: OnCancel]) => R,
   options?: RetryOptions
-): (arg1: P1, arg2: P2, arg3: P3) => R
-
-declare function pretry<R, P1, P2, P3, P4>(
-  fn: (arg1: P1, arg2: P2, arg3: P3, arg4: P4, onCancel?: OnCancel) => R,
-  options?: RetryOptions
-): (arg1: P1, arg2: P2, arg3: P3, arg4: P4) => R
-
-declare function pretry<R, P1, P2, P3, P4, P5>(
-  fn: (
-    arg1: P1,
-    arg2: P2,
-    arg3: P3,
-    arg4: P4,
-    arg5: P5,
-    onCancel?: OnCancel
-  ) => R,
-  options?: RetryOptions
-): (arg1: P1, arg2: P2, arg3: P3, arg4: P4, arg5: P5) => R
-
-declare function pretry<R, P1, P2, P3, P4, P5, P6>(
-  fn: (
-    arg1: P1,
-    arg2: P2,
-    arg3: P3,
-    arg4: P4,
-    arg5: P5,
-    arg6: P6,
-    onCancel?: OnCancel
-  ) => R,
-  options?: RetryOptions
-): (arg1: P1, arg2: P2, arg3: P3, arg4: P4, arg5: P5, arg6: P6) => R
-
-export default pretry
+): (...args: T) => EnsurePromise<R>